Allows FAISS_HNSW_FLAT index to load & search on hnswlib flat index
