+++ include/knowhere/expected.h
@@ -46,6 +46,8 @@ enum class Status {
     invalid_cluster_error = 24,
     cluster_inner_error = 25,
     timeout = 26,
+    internal_error = 27,
+    invalid_serialized_index_type = 28,
 };
 
 inline std::string
@@ -95,6 +97,10 @@ Status2String(knowhere::Status status) {
             return "invalid cluster type";
         case knowhere::Status::cluster_inner_error:
             return "cluster inner error";
+        case knowhere::Status::internal_error:
+            return "internal error (something that must not have happened at all)";
+        case knowhere::Status::invalid_serialized_index_type:
+            return "the serialized index type is not recognized";
         default:
             return "unexpected status";
     }
